+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Bump AndroidX Camera to v1.3.4
 - Bump AndroidX Fragment KTX to v1.8.0
 - Listen to observable medical history field in medical history summary view
+- Fix diagnosis warning dialog showing multiple times on save/back click in summary screen
 
 ### Features